+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ##### Unreleased
 - Added a workaround of a Closure Compiler unsafe optimization, [#972](https://github.com/zloirock/core-js/issues/972)
 - One more fix crashing of `Object.create(null)` on WSH, [#970](https://github.com/zloirock/core-js/issues/970)
+- Added Deno 1.14 compat data mapping
 
 ##### 3.16.1 - 2021.08.09
 - Fixed microtask implementation on iOS Pebble, [#967](https://github.com/zloirock/core-js/issues/967)